- non stable - Added new parameter handling between XML-Parser and Command-objects.
git-svn-id: svn://localhost:8080/caleydo_main/trunk@239 1f7349ae-fd9f-0d40-aeb8-9798e6c0fce3

+++ src/cerberus/manager/data/ICollectionManager.java
@@ -119,7 +119,7 @@ public abstract class ICollectionManager
 
 	protected int getIndexInVector_byUniqueId( final int iItemId ) {
 		Integer buffer = hashId2IndexLookup.get( new Integer( iItemId ));
-		return (int) buffer.intValue();
+		return buffer.intValue();
 		
 		//return (int) hashId2IndexLookup.get( new Integer( iItemId )).intValue();
 	}
